+++ apps/rush-lib/src/cli/actions/BaseInstallAction.ts
@@ -13,6 +13,7 @@ import { PurgeManager } from '../../logic/PurgeManager';
 import { SetupChecks } from '../../logic/SetupChecks';
 import { StandardScriptUpdater } from '../../logic/StandardScriptUpdater';
 import { Stopwatch } from '../../utilities/Stopwatch';
+import { VersionMismatchFinder } from '../../api/VersionMismatchFinder';
 
 /**
  * This is the common base class for InstallAction and UpdateAction.
@@ -58,7 +59,7 @@ export abstract class BaseInstallAction extends BaseRushAction {
 
   protected run(): Promise<void> {
 
-    this.runRushCheckIfNecessary();
+    VersionMismatchFinder.runRushCheckIfNecessary(this.rushConfiguration);
 
     const stopwatch: Stopwatch = Stopwatch.start();
